Implements composer parser
